python-bridge/requirements.txt
MetaTrader5==5.0.45
pandas==2.1.1
numpy==1.26.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
fastapi==0.104.0
uvicorn==0.23.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
schedule==1.2.1